Rename Integration test directory to Feature
